+++ CHANGES.rst
@@ -3,8 +3,9 @@ Version 2.1.3
 
 Unreleased
 
+-   Implement ``format_map``, ``casefold``, ``removeprefix``, and ``removesuffix``
+    methods. :issue:`370`
 -   Fix static typing for basic ``str`` methods on ``Markup``. :issue:`358`
--   Implement ``format_map`` method. :issue:`370`
 
 
 Version 2.1.2